backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic[email]>=2.7.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
anthropic>=0.25.0
rapidfuzz>=3.6.0
# Optional: For PostgreSQL/Neon database (uncomment if needed)
# psycopg2-binary>=2.9.9

# Testing
pytest>=8.0.0
httpx>=0.27.0
pytest-cov>=4.1.0
